requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-jsonapi>=3.2.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
gunicorn==20.0.4
inflection==0.5.1
now==0.0.6
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
